chore: clean up unused dependencies and optimize imports
Apexcharts needs dynamic import

Remove unused dependencies from pnpm-lock.yaml to reduce bloat and 
optimize the project. Update import statements in the codebase to 
use dynamic imports for better performance and to ensure server-side 
rendering compatibility.
