+++ package.json
@@ -17,7 +17,6 @@
     "clsx": "2.1.1",
     "culori": "4.0.1",
     "fast-glob": "3.3.3",
-    "flowbite-react": "0.10.2",
     "lodash.debounce": "4.0.8",
     "lucide-react": "0.475.0",
     "next": "15.1.7",
