+++ src/app/sleep-data/get-data.ts
@@ -1,5 +1,3 @@
-"use server";
-
 import { promises as fs } from "fs";
 import path from "path";
