+++ src/lib/colors.ts
@@ -1,0 +1,7 @@
+export const chartColors = {
+  events: "#da1f3d",
+  hours: "#1788c3",
+  leak: "#ed6937",
+  mask: "#6e308f",
+  score: "#ffc344",
+};
